[frontendActions] Rename package "common" to "backgroundables"
